+++ dags/historical_games_ingestion.py
@@ -1,63 +1,104 @@
+########################################
+# Not using this DAG for now
+########################################
+
 from airflow import DAG
 from airflow.operators.python import PythonOperator
+
+# from airflow.operators.bash import BashOperator
+# from airflow.providers.amazon.aws.transfers.local_to_s3 import (
+#     LocalFilesystemToS3Operator,
+# )
+import boto3
+
+import asyncio
 from datetime import datetime, timedelta
 import json
 import os
-import boto3
-from betflow.api_connectors import ESPNConnector
+import aiohttp
+from betflow.historical.hist_api_connectors import NBAHistoricalConnector
 from betflow.historical.config import HistoricalConfig
+from dotenv import load_dotenv
+
+load_dotenv()
 
 default_args = {
     "owner": "airflow",
-    "depends_on_past": True,
-    "start_date": datetime(2024, 11, 29),
-    "end_date": datetime(2024, 12, 1),
+    "depends_on_past": True,  # Enable sequential processing
+    "start_date": datetime(2024, 11, 27),  # NBA season 2025 start date is 2024-10-22
+    "end_date": datetime(2024, 12, 1),  # Today's date
     "email_on_failure": False,
     "retries": 0,
     "retry_delay": timedelta(minutes=5),
+    # "dagrun_timeout": timedelta(minutes=60),  # Add timeout
     "tags": ["nba", "historical", "games"],  # Add tags
 }
 
 
 def fetch_games_by_date(**context):
-    """Fetch games for a specific date"""
-    logical_date = context.get("data_interval_start")
-    date_str = logical_date.strftime("%Y%m%d")  # Format: YYYYMMDD for ESPN API
+    async def _fetch():
+        """Fetch games and their statistics for a specific date"""
+        logical_date = context.get("data_interval_start", context.get("logical_date"))
+        if not logical_date:
+            raise ValueError("No execution date found in context")
+        date_str = logical_date.strftime("%Y-%m-%d")
 
-    try:
-        espn_connector = ESPNConnector(historical=True)
-        endpoint = "basketball/nba/scoreboard"
-        params = {"dates": date_str}
+        async with aiohttp.ClientSession() as session:
+            games_connector = NBAHistoricalConnector(
+                api_key=os.getenv("API_SPORTS_IO_KEY")
+            )
 
-        # Make rate-limited request
-        raw_data = espn_connector.make_request(endpoint, params)
+            # First fetch games for the date
+            conn_response = await games_connector.fetch_games_by_date(
+                session, "standard", 2024, date_str
+            )
+            games_data = conn_response["response"]
+            complete_games_data = []
+            for game in games_data:
+                game_id = str(game["id"])
+                game_stats = await games_connector.fetch_game_statistics(
+                    session, game_id
+                )
 
-        # Process games using existing transformer
-        processed_games = []
-        for game in raw_data.get("events", []):
-            game_data = espn_connector.api_raw_nba_data(game)
-            if game_data["status_state"] in ["post", "STATUS_FINAL"]:
-                processed_games.append(game_data)
+                # Process and combine game and statistics data
+                processed_game = games_connector.process_game_data(game, game_stats)
+                complete_games_data.append(processed_game)
 
-        if processed_games:
-            # Write to temporary location
-            output_dir = f"/tmp/{HistoricalConfig.S3_PATHS['games_prefix']}/{logical_date.strftime('%Y-%m-%d')}"
-            os.makedirs(output_dir, exist_ok=True)
+            if complete_games_data:
+                context["task_instance"].xcom_push(
+                    key="games_data", value=complete_games_data
+                )
+                context["task_instance"].xcom_push(key="date", value=date_str)
 
-            with open(f"{output_dir}/games.json", "w") as f:
-                json.dump(processed_games, f)
+            return len(complete_games_data)
 
-            context["task_instance"].xcom_push(key="games_data", value=processed_games)
-            return len(processed_games)
-        return 0
+    # Run async code in sync context
+    return asyncio.run(_fetch())
 
-    except Exception as e:
-        print(f"Error fetching games for date {date_str}: {str(e)}")
-        raise
+
+def process_games_data(**context):
+    """Store processed games data to S3"""
+    games_data = context["task_instance"].xcom_pull(key="games_data")
+    date_str = context["task_instance"].xcom_pull(key="date")
+
+    if not games_data:
+        print("No games data found for processing")
+        return None
+
+    # Use config for output paths
+    output_dir = f"/tmp/{HistoricalConfig.S3_PATHS['games_prefix']}/{date_str}"
+    os.makedirs(output_dir, exist_ok=True)
+
+    with open(f"{output_dir}/games.json", "w") as f:
+        json.dump(games_data, f)
+
+    # for historical odds DAG
+    # context["task_instance"].xcom_push(key="games_data", value=games_data)
+
+    return output_dir
 
 
 def upload_to_s3_func(**context):
-    """Upload processed games data to S3"""
     date_str = context["ds"]
     local_path = (
         f"/tmp/{HistoricalConfig.S3_PATHS['games_prefix']}/{date_str}/games.json"
@@ -67,7 +108,12 @@ def upload_to_s3_func(**context):
     s3_client = boto3.client("s3")
     try:
         s3_client.upload_file(
-            local_path, HistoricalConfig.S3_PATHS["raw_bucket"], s3_path
+            local_path,
+            HistoricalConfig.S3_PATHS["raw_bucket"],
+            s3_path,
+            ExtraArgs={
+                "ServerSideEncryption": "AES256"
+            },  # Add encryption if bucket requires it
         )
     except Exception as e:
         print(f"Error uploading to S3: {str(e)}")
@@ -77,9 +123,9 @@ def upload_to_s3_func(**context):
 with DAG(
     "historical_games_ingestion",
     default_args=default_args,
-    description="Ingest historical NBA games data with backfill support",
+    description="Ingest historical games data with backfill support",
     schedule_interval="@daily",
-    catchup=True,
+    catchup=True,  # Enable backfilling
 ) as dag:
     fetch_daily_games = PythonOperator(
         task_id="fetch_daily_games",
@@ -87,8 +133,27 @@ with DAG(
         provide_context=True,
     )
 
+    process_daily_games = PythonOperator(
+        task_id="process_daily_games",
+        python_callable=process_games_data,
+        provide_context=True,
+    )
+
+    # upload_to_s3 = LocalFilesystemToS3Operator(
+    #     task_id="upload_to_s3",
+    #     filename=f"/tmp/{HistoricalConfig.S3_PATHS['games_prefix']}/{{{{ ds }}}}/games.json",
+    #     dest_key=f"{HistoricalConfig.S3_PATHS['games_prefix']}/nba/{{{{ ds }}}}/games.json",
+    #     dest_bucket=HistoricalConfig.S3_PATHS["raw_bucket"],
+    #     # aws_conn_id=None,  # "aws_default",
+    #     replace=True,
+    # )
+    # upload_to_s3 = BashOperator(
+    #     task_id="upload_to_s3",
+    #     bash_command=f'aws s3 sync /tmp/{HistoricalConfig.S3_PATHS["games_prefix"]}/{{{{ ds }}}}/games.json '
+    #     f's3://{HistoricalConfig.S3_PATHS["raw_bucket"]}/{HistoricalConfig.S3_PATHS["games_prefix"]}/nba/{{{{ ds }}}}/games.json',
+    # )
     upload_to_s3 = PythonOperator(
         task_id="upload_to_s3", python_callable=upload_to_s3_func, provide_context=True
     )
 
-    fetch_daily_games >> upload_to_s3
+    fetch_daily_games >> process_daily_games >> upload_to_s3
