+++ src/betflow/historical/config/historical_config.py
@@ -1,3 +1,6 @@
+from datetime import datetime
+
+
 class HistoricalConfig:
     # Game duration in minutes
     GAME_DURATIONS = {
@@ -30,3 +33,26 @@ class HistoricalConfig:
 
     # League IDs
     LEAGUE_IDS = {"nfl": 1, "ncaa": 2, "nba": "standard", "nhl": 57}
+
+    SPORT_CONFIGS = {
+        "nba": {
+            "endpoint": "basketball/nba/scoreboard",
+            "start_date": datetime(2024, 10, 22),  # NBA season start
+            "end_date": datetime(2024, 12, 1),
+        },
+        "nfl": {
+            "endpoint": "football/nfl/scoreboard",
+            "start_date": datetime(2024, 9, 7),  # NFL season start
+            "end_date": datetime(2024, 12, 1),
+        },
+        "nhl": {
+            "endpoint": "hockey/nhl/scoreboard",
+            "start_date": datetime(2024, 10, 10),  # NHL season start
+            "end_date": datetime(2024, 12, 1),
+        },
+        "cfb": {
+            "endpoint": "football/college-football/scoreboard",
+            "start_date": datetime(2024, 8, 26),  # CFB season start
+            "end_date": datetime(2024, 12, 1),
+        },
+    }
