+++ src/betflow/historical/hist_api_connectors/__init__.py
@@ -4,10 +4,12 @@ from .hist_game_conn import (
     FootballHistoricalConnector,
 )
 from .hist_odds_conn import HistoricalOddsConnector
+from .batch_game_connector import ESPNBatchConnector
 
 __all__ = [
     "NBAHistoricalConnector",
     "NHLHistoricalConnector",
     "FootballHistoricalConnector",
     "HistoricalOddsConnector",
+    "ESPNBatchConnector",
 ]
